+++ sdk/resourcemanager/containerservice/armcontainerservice/models_serde.go
@@ -214,6 +214,33 @@ func (a *AgentPoolAvailableVersionsPropertiesAgentPoolVersionsItem) UnmarshalJSO
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type AgentPoolGPUProfile.
+func (a AgentPoolGPUProfile) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "installGPUDriver", a.InstallGPUDriver)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AgentPoolGPUProfile.
+func (a *AgentPoolGPUProfile) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "installGPUDriver":
+			err = unpopulate(val, "InstallGPUDriver", &a.InstallGPUDriver)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type AgentPoolListResult.
 func (a AgentPoolListResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -1106,6 +1133,7 @@ func (i *IstioCertificateAuthority) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type IstioComponents.
 func (i IstioComponents) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
+	populate(objectMap, "egressGateways", i.EgressGateways)
 	populate(objectMap, "ingressGateways", i.IngressGateways)
 	return json.Marshal(objectMap)
 }
@@ -1119,8 +1147,42 @@ func (i *IstioComponents) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
+		case "egressGateways":
+			err = unpopulate(val, "EgressGateways", &i.EgressGateways)
+			delete(rawMsg, key)
 		case "ingressGateways":
 			err = unpopulate(val, "IngressGateways", &i.IngressGateways)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", i, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type IstioEgressGateway.
+func (i IstioEgressGateway) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "enabled", i.Enabled)
+	populate(objectMap, "nodeSelector", i.NodeSelector)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type IstioEgressGateway.
+func (i *IstioEgressGateway) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", i, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "enabled":
+			err = unpopulate(val, "Enabled", &i.Enabled)
+			delete(rawMsg, key)
+		case "nodeSelector":
+			err = unpopulate(val, "NodeSelector", &i.NodeSelector)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -2112,6 +2174,7 @@ func (m ManagedClusterAgentPoolProfile) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "enableNodePublicIP", m.EnableNodePublicIP)
 	populate(objectMap, "enableUltraSSD", m.EnableUltraSSD)
 	populate(objectMap, "gpuInstanceProfile", m.GpuInstanceProfile)
+	populate(objectMap, "gpuProfile", m.GpuProfile)
 	populate(objectMap, "hostGroupID", m.HostGroupID)
 	populate(objectMap, "kubeletConfig", m.KubeletConfig)
 	populate(objectMap, "kubeletDiskType", m.KubeletDiskType)
@@ -2195,6 +2258,9 @@ func (m *ManagedClusterAgentPoolProfile) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "gpuInstanceProfile":
 			err = unpopulate(val, "GpuInstanceProfile", &m.GpuInstanceProfile)
+			delete(rawMsg, key)
+		case "gpuProfile":
+			err = unpopulate(val, "GpuProfile", &m.GpuProfile)
 			delete(rawMsg, key)
 		case "hostGroupID":
 			err = unpopulate(val, "HostGroupID", &m.HostGroupID)
@@ -2327,6 +2393,7 @@ func (m ManagedClusterAgentPoolProfileProperties) MarshalJSON() ([]byte, error) 
 	populate(objectMap, "enableNodePublicIP", m.EnableNodePublicIP)
 	populate(objectMap, "enableUltraSSD", m.EnableUltraSSD)
 	populate(objectMap, "gpuInstanceProfile", m.GpuInstanceProfile)
+	populate(objectMap, "gpuProfile", m.GpuProfile)
 	populate(objectMap, "hostGroupID", m.HostGroupID)
 	populate(objectMap, "kubeletConfig", m.KubeletConfig)
 	populate(objectMap, "kubeletDiskType", m.KubeletDiskType)
@@ -2409,6 +2476,9 @@ func (m *ManagedClusterAgentPoolProfileProperties) UnmarshalJSON(data []byte) er
 			delete(rawMsg, key)
 		case "gpuInstanceProfile":
 			err = unpopulate(val, "GpuInstanceProfile", &m.GpuInstanceProfile)
+			delete(rawMsg, key)
+		case "gpuProfile":
+			err = unpopulate(val, "GpuProfile", &m.GpuProfile)
 			delete(rawMsg, key)
 		case "hostGroupID":
 			err = unpopulate(val, "HostGroupID", &m.HostGroupID)
@@ -3782,7 +3852,11 @@ func (m *ManagedClusterProperties) UnmarshalJSON(data []byte) error {
 func (m ManagedClusterPropertiesAutoScalerProfile) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "balance-similar-node-groups", m.BalanceSimilarNodeGroups)
+	populate(objectMap, "daemonset-eviction-for-empty-nodes", m.DaemonsetEvictionForEmptyNodes)
+	populate(objectMap, "daemonset-eviction-for-occupied-nodes", m.DaemonsetEvictionForOccupiedNodes)
 	populate(objectMap, "expander", m.Expander)
+	populate(objectMap, "expanders", m.Expanders)
+	populate(objectMap, "ignore-daemonsets-utilization", m.IgnoreDaemonsetsUtilization)
 	populate(objectMap, "max-empty-bulk-delete", m.MaxEmptyBulkDelete)
 	populate(objectMap, "max-graceful-termination-sec", m.MaxGracefulTerminationSec)
 	populate(objectMap, "max-node-provision-time", m.MaxNodeProvisionTime)
@@ -3813,8 +3887,20 @@ func (m *ManagedClusterPropertiesAutoScalerProfile) UnmarshalJSON(data []byte) e
 		case "balance-similar-node-groups":
 			err = unpopulate(val, "BalanceSimilarNodeGroups", &m.BalanceSimilarNodeGroups)
 			delete(rawMsg, key)
+		case "daemonset-eviction-for-empty-nodes":
+			err = unpopulate(val, "DaemonsetEvictionForEmptyNodes", &m.DaemonsetEvictionForEmptyNodes)
+			delete(rawMsg, key)
+		case "daemonset-eviction-for-occupied-nodes":
+			err = unpopulate(val, "DaemonsetEvictionForOccupiedNodes", &m.DaemonsetEvictionForOccupiedNodes)
+			delete(rawMsg, key)
 		case "expander":
 			err = unpopulate(val, "Expander", &m.Expander)
+			delete(rawMsg, key)
+		case "expanders":
+			err = unpopulate(val, "Expanders", &m.Expanders)
+			delete(rawMsg, key)
+		case "ignore-daemonsets-utilization":
+			err = unpopulate(val, "IgnoreDaemonsetsUtilization", &m.IgnoreDaemonsetsUtilization)
 			delete(rawMsg, key)
 		case "max-empty-bulk-delete":
 			err = unpopulate(val, "MaxEmptyBulkDelete", &m.MaxEmptyBulkDelete)
@@ -4633,6 +4719,7 @@ func (m *ManagedClusterWorkloadAutoScalerProfileKeda) UnmarshalJSON(data []byte)
 // MarshalJSON implements the json.Marshaller interface for type ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler.
 func (m ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
+	populate(objectMap, "addonAutoscaling", m.AddonAutoscaling)
 	populate(objectMap, "enabled", m.Enabled)
 	return json.Marshal(objectMap)
 }
@@ -4646,6 +4733,9 @@ func (m *ManagedClusterWorkloadAutoScalerProfileVerticalPodAutoscaler) Unmarshal
 	for key, val := range rawMsg {
 		var err error
 		switch key {
+		case "addonAutoscaling":
+			err = unpopulate(val, "AddonAutoscaling", &m.AddonAutoscaling)
+			delete(rawMsg, key)
 		case "enabled":
 			err = unpopulate(val, "Enabled", &m.Enabled)
 			delete(rawMsg, key)